# middle_platform/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /demo:
        get:
            tags:
                - Demo
            operationId: Demo_GetDemo
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.helloworld.v1.GetDemoReply'
    /exchange-rates:
        get:
            tags:
                - ExchangeRate
            operationId: ExchangeRate_BaseCurrency
            parameters:
                - name: base
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/exchange_rate.v1.BaseCurrencyReply'
    /helloworld/{name}:
        get:
            tags:
                - Greeter
            description: Sends a greeting
            operationId: Greeter_SayHello
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.HelloReply'
    /supported-currencies:
        get:
            tags:
                - ExchangeRate
            description: get Supported Currencies
            operationId: ExchangeRate_SupportedCurrencies
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/exchange_rate.v1.RateReply'
    /v1/nfts/mute:
        post:
            tags:
                - NftTransfer
            operationId: NftTransfer_PostReportAccountMute
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.nft_transfer.v1.PostReportAccountMuteRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.nft_transfer.v1.PostReportAccountMuteReply'
    /v1/nfts/report/spam:
        get:
            tags:
                - NftTransfer
            operationId: NftTransfer_GetReportSpam
            parameters:
                - name: collectionId
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: string
                - name: page
                  in: query
                  description: uint32 cursor = 3;
                  schema:
                    type: integer
                    format: uint32
                - name: page_size
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: source
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.nft_transfer.v1.GetReportSpamReply'
        post:
            tags:
                - NftTransfer
            operationId: NftTransfer_PostReportSpam
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.nft_transfer.v1.PostReportSpamRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.nft_transfer.v1.PostReportSpamReply'
    /v1/nfts/transfer:
        post:
            tags:
                - NftTransfer
            operationId: NftTransfer_GetTransferNft
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.nft_transfer.v1.GetTransferNftRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.nft_transfer.v1.GetTransferNftReply'
    /v1/nfttransfer/getnfttransfer:
        get:
            tags:
                - NftTransfer
            operationId: NftTransfer_GetNftTransfer
            parameters:
                - name: address
                  in: query
                  schema:
                    type: string
                - name: network
                  in: query
                  schema:
                    type: string
                - name: type
                  in: query
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  schema:
                    type: string
                - name: orderDirection
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: uint64
                - name: cursor
                  in: query
                  schema:
                    type: integer
                    format: uint64
                - name: accountId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.nft_transfer.v1.GetNftTransferReply'
components:
    schemas:
        api.helloworld.v1.GetDemoReply:
            type: object
            properties: {}
        api.nft_transfer.v1.AccountMuteReport:
            type: object
            properties:
                accountId:
                    type: string
                collectionId:
                    type: string
                createdAt:
                    type: string
        api.nft_transfer.v1.ActionStArr:
            type: object
            properties:
                tag:
                    type: string
                type:
                    type: string
                index:
                    type: integer
                    format: uint32
                address_from:
                    type: string
                address_to:
                    type: string
                cost:
                    $ref: '#/components/schemas/api.nft_transfer.v1.CostSt'
                contract_address:
                    type: string
                token_id:
                    type: string
        api.nft_transfer.v1.CostSt:
            type: object
            properties:
                symbol:
                    type: string
                value:
                    type: string
                decimals:
                    type: integer
                    format: uint32
        api.nft_transfer.v1.GetNftTransferReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                reason:
                    type: string
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/api.nft_transfer.v1.PnftTransferSt'
        api.nft_transfer.v1.GetReportSpamReply:
            type: object
            properties:
                code:
                    type: integer
                    format: uint32
                page:
                    type: integer
                    description: optional uint32 cursor = 2;
                    format: uint32
                page_size:
                    type: integer
                    format: uint32
                total:
                    type: integer
                    format: uint64
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.nft_transfer.v1.SpamReport'
        api.nft_transfer.v1.GetTransferNftReply:
            type: object
            properties:
                code:
                    type: integer
                    format: uint32
                page:
                    type: integer
                    format: uint32
                page_size:
                    type: integer
                    format: uint32
                total:
                    type: integer
                    format: uint64
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.nft_transfer.v1.TransferNft'
        api.nft_transfer.v1.GetTransferNftRequest:
            type: object
            properties:
                contractAddress:
                    type: string
                eventType:
                    type: string
                owners:
                    type: string
                page:
                    type: integer
                    format: uint32
                page_size:
                    type: integer
                    format: uint32
        api.nft_transfer.v1.NodeStArr:
            type: object
            properties:
                timestamp:
                    type: string
                hash:
                    type: string
                owner:
                    type: string
                address_from:
                    type: string
                address_to:
                    type: string
                network:
                    type: string
                tag:
                    type: string
                type:
                    type: string
                actions:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.nft_transfer.v1.ActionStArr'
        api.nft_transfer.v1.PnftTransferSt:
            type: object
            properties:
                cursor:
                    type: string
                result:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.nft_transfer.v1.NodeStArr'
        api.nft_transfer.v1.PostReportAccountMuteReply:
            type: object
            properties:
                code:
                    type: integer
                    format: uint32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/api.nft_transfer.v1.AccountMuteReport'
        api.nft_transfer.v1.PostReportAccountMuteRequest:
            type: object
            properties:
                accountId:
                    type: string
                collectionId:
                    type: string
        api.nft_transfer.v1.PostReportSpamReply:
            type: object
            properties:
                code:
                    type: integer
                    format: uint32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/api.nft_transfer.v1.SpamReport'
        api.nft_transfer.v1.PostReportSpamRequest:
            type: object
            properties:
                collectionId:
                    type: string
                status:
                    type: string
                createAt:
                    type: string
                updateAt:
                    type: string
                source:
                    type: string
        api.nft_transfer.v1.SpamReport:
            type: object
            properties:
                collection_id:
                    type: string
                status:
                    type: string
                create_at:
                    type: string
                update_at:
                    type: string
                source:
                    type: string
        api.nft_transfer.v1.TransferNft:
            type: object
            properties:
                contractAddress:
                    type: string
                nftId:
                    type: string
                chain:
                    type: string
                tokenId:
                    type: string
                collectionId:
                    type: string
                eventType:
                    type: string
                addressFrom:
                    type: string
                addressTo:
                    type: string
                owner:
                    type: string
                blockTimestamp:
                    type: string
        exchange_rate.v1.BaseCurrencyReply:
            type: object
            properties:
                timestamp:
                    type: integer
                    format: int32
                base:
                    type: string
                rates:
                    type: object
                    additionalProperties:
                        type: number
                        format: double
        exchange_rate.v1.RateReply:
            type: object
            properties:
                currencies:
                    type: array
                    items:
                        type: string
        helloworld.v1.HelloReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the greetings
tags:
    - name: Demo
    - name: ExchangeRate
      description: The rate service definition.
    - name: Greeter
      description: The greeting service definition.
    - name: NftTransfer